pyproject.toml
[tool.poetry]
name = "repl_nix_discord-one-day-one-zenn"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
python-dotenv = "^0.21.1"
requests = "^2.28.2"
replit = "^3.2.5"
aioflask = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
